refactor persistor file structure
