+++ src/store/persistor.ts
@@ -1,0 +1,6 @@
+import { persistStore } from 'redux-persist';
+import store from './store';
+
+const persistor = persistStore(store);
+
+export default persistor;
